Version: 2012-05-31 ksh93u+
Sourced from ast-open-history version 2012-05-31 and:
7b97b44798d6fa664434b53f0f5e48b98103f25108c1d03e7d58dedf2ab7af0d  INIT.2012-05-31.tgz

+++ src/lib/libast/comp/iconv.c
@@ -423,7 +423,7 @@ _ast_iconv_name(register const char* m, register char* b, size_t n)
 	const _ast_iconv_list_t*		bp;
 	register int				c;
 	register char*				e;
-	int					sub[2];
+	ssize_t					sub[2];
 	char					buf[16];
 #if DEBUG_TRACE
 	char*					o;
